+++ src/api/consumer/lib/test-inline.js
@@ -3,14 +3,14 @@ import { loadConsumer } from '../../../consumer';
 import InlineId from '../../../consumer/bit-inline-id';
 import Bit from '../../../consumer/component';
 
-export function testInline(id: string, verbose: boolean): Promise<Bit> {
+export function testInline(id: string, verbose: boolean = true): Promise<Bit> {
   return loadConsumer()
     .then(consumer =>
       consumer.runComponentSpecs(InlineId.parse(id), verbose)
     );
 }
 
-export function testInlineAll(verbose: boolean): Promise<Bit> {
+export function testInlineAll(verbose: boolean = true): Promise<Bit> {
   return loadConsumer()
     .then(consumer =>
       consumer.runAllInlineSpecs(verbose)