+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- fix the "--verbose" flag of `bit test`
 - add logger
 - support debug-mode for e2e tests